Backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dj-database-url==2.1.0
geographiclib==2.1
geopy==2.4.1
numpy==1.26.4
pandas==3.0.1
pgvector==0.2.4
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
tensorflow==2.16.1
tzdata==2025.3

